% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-git_html_calendar.R
\name{git_html_calendar}
\alias{git_html_calendar}
\title{Get all projects from Github Organisation}
\usage{
git_html_calendar(projects)
}
\arguments{
\item{token}{token with access to repos and projects}

\item{org}{Organisation name}
}
\value{
Projects
}
\description{
Obtains list from Github Organisation
}
