% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-git_issues.R
\name{git_issues}
\alias{git_issues}
\title{Get all projects from Github Organisation}
\usage{
git_issues(projects)
}
\arguments{
\item{token}{token with access to repos and projects}

\item{org}{Organisation name}
}
\value{
Projects
}
\description{
Obtains list from Github Organisation
}
