% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-git_projects.R
\name{git_projects}
\alias{git_projects}
\title{Get all projects from Github Organisation}
\usage{
git_projects(token, org)
}
\arguments{
\item{token}{token with access to repos and projects}

\item{org}{Organisation name}
}
\value{
Projects
}
\description{
Obtains list from Github Organisation
}
