% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-git_schedule.R
\name{git_schedule}
\alias{git_schedule}
\title{Get all projects from Github Organisation}
\usage{
git_schedule(token, org)
}
\arguments{
\item{token}{token with access to repos and projects}

\item{org}{Organisation name}
}
\value{
Projects
}
\description{
Obtains list from Github Organisation
}
